Version 3.4.0

+++ Project/STM32F10x_StdPeriph_Examples/I2S/Interrupt/stm32f10x_it.c
@@ -2,8 +2,8 @@
   ******************************************************************************
   * @file    I2S/Interrupt/stm32f10x_it.c 
   * @author  MCD Application Team
-  * @version V3.3.0
-  * @date    04/16/2010
+  * @version V3.4.0
+  * @date    10/15/2010
   * @brief   Main Interrupt Service Routines.
   *          This file provides template for all exceptions handler and peripherals
   *          interrupt service routine.
@@ -37,7 +37,7 @@
 /* Private variables ---------------------------------------------------------*/
 extern const uint16_t I2S3_Buffer_Tx[32];
 extern uint16_t I2S2_Buffer_Rx[32];
-extern __IO uint8_t TxIdx, RxIdx;
+extern  __IO uint32_t TxIdx, RxIdx;
 
 /* Private function prototypes -----------------------------------------------*/
 /* Private functions ---------------------------------------------------------*/
