+++ Project/STM32F10x_StdPeriph_Examples/USART/HalfDuplex/main.c
@@ -1,9 +1,9 @@
 /**
   ******************************************************************************
-  * @file USART/HalfDuplex/main.c 
+  * @file    USART/HalfDuplex/main.c 
   * @author  MCD Application Team
-  * @version V3.3.0
-  * @date    04/16/2010
+  * @version V3.4.0
+  * @date    10/15/2010
   * @brief   Main program body
   ******************************************************************************
   * @copy
@@ -46,8 +46,8 @@ uint8_t TxBuffer1[] = "USART Half Duplex: USARTy -> USARTz using HalfDuplex mode
 uint8_t TxBuffer2[] = "USART Half Duplex: USARTz -> USARTy using HalfDuplex mode";
 uint8_t RxBuffer1[TxBufferSize2];
 uint8_t RxBuffer2[TxBufferSize1];
-uint8_t NbrOfDataToRead1 = TxBufferSize2;
-uint8_t NbrOfDataToRead2 = TxBufferSize1;
+uint32_t NbrOfDataToRead1 = TxBufferSize2;
+uint32_t NbrOfDataToRead2 = TxBufferSize1;
 uint8_t TxCounter1 = 0, RxCounter1 = 0;
 uint8_t TxCounter2 = 0, RxCounter2 = 0;
 volatile TestStatus TransferStatus1 = FAILED, TransferStatus2 = FAILED;
